allow VoiceActivityVideoSampler to be disabled with 0fps
also simplified realtime_video_agent example
